Initial Commit of generic Package
The generic package contains Contains an interface and implementation
for a generic tree. A generic tree mirrors a tree, but allows for
generic values at each node rather than simply integers. Generic trees
can be interacted with using the values contained in the tree, the
integer IDs for these values, or a combination of both.

This commit includes implementations and comments.